Handle multiple Python packages

+++ jupyter_releaser/python.py
@@ -17,16 +17,17 @@ PYPROJECT = util.PYPROJECT
 SETUP_PY = util.SETUP_PY
 
 
-def build_dist(dist_dir):
+def build_dist(dist_dir, clean=True):
     """Build the python dist files into a dist folder"""
     # Clean the dist folder of existing npm tarballs
     os.makedirs(dist_dir, exist_ok=True)
     dest = Path(dist_dir)
-    for pkg in glob(f"{dest}/*.gz") + glob(f"{dest}/*.whl"):
-        os.remove(pkg)
+    if clean:
+        for pkg in glob(f"{dest}/*.gz") + glob(f"{dest}/*.whl"):
+            os.remove(pkg)
 
     if PYPROJECT.exists():
-        util.run(f"python -m build --outdir {dest} .", quiet=True)
+        util.run(f"python -m build --outdir {dest} .", quiet=True, show_cwd=True)
     elif SETUP_PY.exists():
         util.run(f"python setup.py sdist --dist-dir {dest}", quiet=True)
         util.run(f"python setup.py bdist_wheel --dist-dir {dest}", quiet=True)
@@ -60,7 +61,7 @@ def check_dist(dist_file, test_cmd=""):
         util.run(f"{bin_path}/{test_cmd}")
 
 
-def get_pypi_token(release_url):
+def get_pypi_token(release_url, python_package):
     """Get the PyPI token
 
     Note: Do not print the token in CI since it will not be sanitized
@@ -70,6 +71,8 @@ def get_pypi_token(release_url):
     if pypi_token_map and release_url:
         parts = release_url.replace("https://github.com/", "").split("/")
         repo_name = f"{parts[0]}/{parts[1]}"
+        if python_package != ".":
+            repo_name += f"/{python_package}"
         util.log(f"Looking for PYPI token for {repo_name} in token map")
         for line in pypi_token_map.splitlines():
             name, _, token = line.partition(",")
